requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
openai==1.12.0
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
pytest==8.0.0
pytest-asyncio==0.23.5
supabase
fal-client
